Updates (#6)
* Adding USER root to the dockerfile

* Fixing deployment assignment in JenkinsfileLab

The deployment variable assignment in the final step was failing due to the variable RELEASE_NAME not being set. I removed the reference to the release name altogether.

Signed-off-by: Oliver Rodriguez <odrodrig@us.ibm.com>

* Changed context root in server.xml

The context root was not correct and causing issues with ICP. Also added more detail to readme.md

Signed-off-by: Oliver Rodriguez <odrodrig@us.ibm.com>

* Removing Ingress

Reverting ClusterIP to NodePort in the values.yaml files

Signed-off-by: Oliver Rodriguez <odrodrig@us.ibm.com>

* Replacing NodePort with ClusterIP in the Values.yaml

Signed-off-by: Oliver Rodriguez <odrodrig@us.ibm.com>

* Finished instructions

Finished the first draft of instructions. Will continue to add images and additional details.

Signed-off-by: Oliver Rodriguez <odrodrig@us.ibm.com>

* Added readme images

Added the remaining images to the readme that were missing.

Signed-off-by: Oliver Rodriguez <odrodrig@us.ibm.com>